+++ src/Security/UserChecker.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace App\Security;
 
 use App\Entity\User;
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\CustomUserMessageAccountStatusException;
 use Symfony\Component\Security\Core\User\UserCheckerInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -29,7 +30,7 @@ class UserChecker implements UserCheckerInterface
             return;
         }
 
-        if ($user->isAnonymized()) {
+        if ($user->isAnonymized() || null !== $user->getDeletedAt()) {
             throw new CustomUserMessageAccountStatusException('This account has been deactivated.');
         }
 
@@ -38,7 +39,7 @@ class UserChecker implements UserCheckerInterface
         }
     }
 
-    public function checkPostAuth(UserInterface $user): void
+    public function checkPostAuth(UserInterface $user, ?TokenInterface $token = null): void
     {
     }
 }